change workflow divider
