+++ packages/app/src/app/page.tsx
@@ -12,32 +12,8 @@ import frontifyHubTheme from "./theme";
 import { Canvas } from "../../../canvas/src/Canvas";
 
 export default function Home() {
-  const { t, changeLanguage } = useLocalization();
-  const { colorMode, toggleColorMode } = useColorMode();
   return (
     <main className={styles.main}>
-      {t("home.body")}
-      <ColorModeScript initialColorMode={frontifyHubTheme.initialColorMode} />
-      <Button onClick={toggleColorMode}>
-        Toggle {colorMode === "light" ? "Dark" : "Light"}
-      </Button>
-      <button
-        onClick={() => {
-          changeLanguage(Language.Spanish);
-        }}
-      >
-        {t("button.language.spanish")}
-      </button>
-      <button
-        onClick={() => {
-          changeLanguage(Language.English);
-        }}
-      >
-        {t("button.language.english")}
-      </button>
-      <Button bgColor="fron">adasdsadsa</Button>
-      <h1>Home</h1>
-      <Link href="/about">About</Link>
       <Canvas />
     </main>
   );
