--- packages/design-component/src/models/Button.ts
@@ -1,12 +1,0 @@
-export interface IButton {
-    backgroundColor: string,
-    color: string,
-    borderRadius: number,
-    width: number,
-    height: number,
-
-    changeColor: (color: string) => void;
-    changeWidth: (width: number) => void;
-    changeHeight: (height: number) => void;
-    getColor: () => string
-}
