+++ packages/design-component/src/models/IBaseStyle.ts
