--- packages/design-component/src/models/baseStyleCategories/SizingCategory/IHeight.ts
@@ -1,7 +1,0 @@
-import { IBaseStyle } from "../IBaseStyle";
-import { SizeUnit } from "./ISize";
-
-export interface IHeight extends IBaseStyle {
-    height?: SizeUnit
-    type?: 'height' | 'min-height' | 'max-height';
-}
