--- packages/design-component/src/models/baseStyleCategories/SizingCategory/ISize.ts
@@ -1,8 +1,0 @@
-import { AbsoluteSize } from "../../baseUnit/AbsoluteSize";
-import { RelativeSize } from "../../baseUnit/RelativeSize";
-
-export type SizeUnit = AbsoluteSize | RelativeSize;
-
-export interface ISize {
-    setSize(value: string): this;
-}
