--- packages/design-component/src/models/baseStyleCategories/SizingCategory/IWidth.ts
@@ -1,7 +1,0 @@
-import { IBaseStyle } from "../IBaseStyle";
-import { SizeUnit } from "./ISize";
-
-export interface IWidth extends IBaseStyle {
-    width?: SizeUnit;
-    type?: 'width' | 'min-width' | 'max-width'
-}
