--- packages/design-component/src/models/baseStyleCategories/SizingCategory/index.ts
@@ -1,3 +1,0 @@
-export * from './IHeight';
-export * from './IWidth';
-export * from './ISizingCategory';
