--- packages/design-component/src/models/baseStyleCategories/SpacingCategory/IPadding.ts
@@ -1,20 +1,0 @@
-import { BaseStyle } from "../../../view-models/baseStyleCategories/BaseStyle";
-import { AbsoluteSize } from "../../baseUnit/AbsoluteSize";
-import { RelativeSize } from "../../baseUnit/RelativeSize";
-import { IBaseStyle } from "../IBaseStyle";
-
-export type PaddingSize = AbsoluteSize | RelativeSize;
-
-export interface IPadding extends BaseStyle {
-  left?: PaddingSize;
-  right?: PaddingSize;
-  top?: PaddingSize;
-  bottom?: PaddingSize;
-
-  setPadding(value: string): this;
-  setLeft(value: string): this;
-  setRight(value: string): this;
-  setTop(value: string): this;
-  setBottom(value: string): this;
-
-}
