--- packages/design-component/src/models/baseStyleCategories/SpacingCategory/ISpacingCategory.ts
@@ -1,11 +1,0 @@
-import { Margin } from "../../../view-models/baseStyleCategories/SpacingCategory/Margin";
-import { Padding } from "../../../view-models/baseStyleCategories/SpacingCategory/Padding";
-
-export interface ISpacingCategory {
-    padding: Padding;
-    margin: Margin
-
-    setPadding(value: string): this;
-    setMargin(value: string): this;
-
-}
