--- packages/design-component/src/models/baseStyleCategories/SpacingCategory/index.ts
@@ -1,3 +1,0 @@
-export * from './IPadding';
-export * from './IMargin';
-export * from './ISpacingCategory';
