--- packages/design-component/src/models/baseStyleCategories/index.ts
@@ -1,3 +1,0 @@
-export * from './SpacingCategory';
-export * from './IBaseStyle';
-export * from './SizingCategory';
