+++ packages/design-component/src/models/baseUnit/SizeUnit.ts
@@ -1,0 +1,4 @@
+import { AbsoluteSize } from "./AbsoluteSize";
+import { RelativeSize } from "./RelativeSize";
+
+export type SizeUnit = AbsoluteSize | RelativeSize;
