--- packages/design-component/src/models/index.ts
@@ -1,2 +1,0 @@
-export * from "./Button";
-export * from './baseStyleCategories';
