--- packages/design-component/src/tags/index.ts
@@ -1,1 +1,0 @@
-export * from './BoxInstance';
