--- packages/design-component/src/view-models/baseStyleCategories/SizingCategory/index.ts
@@ -1,3 +1,0 @@
-export * from './Height';
-export * from './SizingCategory';
-export * from './Width';
