--- packages/design-component/src/view-models/baseStyleCategories/SpacingCategory/index.ts
@@ -1,1 +1,0 @@
-export * from './SpacingCategory';
