--- packages/design-component/src/view-models/baseStyleCategories/index.ts
@@ -1,2 +1,0 @@
-export * from './SpacingCategory';
-export * from './SizingCategory';
